+++ backend/demo/src/main/java/project_z/demo/dto/TitleDtos/TitlePatchUpdateDto.java
@@ -21,6 +21,6 @@ public class TitlePatchUpdateDto {
     private JsonNullable<String> titleName = JsonNullable.undefined();
     private JsonNullable<TitleStatus> status = JsonNullable.undefined();
     private JsonNullable<Map<String,Float>> rating = JsonNullable.undefined();
-    private JsonNullable<Double> customOrder = JsonNullable.undefined();
+    private JsonNullable<Long> customOrder = JsonNullable.undefined();
     private JsonNullable<String> imageUrl = JsonNullable.undefined();
 }
